+++ Sources/Nuke/Pipeline/ImagePipeline+Configuration.swift
@@ -108,6 +108,10 @@ extension ImagePipeline {
         /// previews have ``ImageContainer/isPreview`` flag set to `true`.
         public var isStoringPreviewsInMemoryCache = true
 
+        /// `false` by default. If `true`, the pipeline will store original images
+        /// in the memory cache when fetching processed images.
+        public var isStoringOriginalImagesInMemoryCache = false
+
         /// If the data task is terminated (either because of a failure or a
         /// cancellation) and the image was partially loaded, the next load will
         /// resume where it left off. Supports both validators (`ETag`,